\name{StatDataML-internal}
\alias{readDimensionSDML}
\alias{writeDimensionSDML}
\alias{readDfSDML}
\alias{catSDML}
\alias{etagsSDML}
\alias{getAttrSDML}
\alias{handlersSDML}
\alias{writeDatasetSDML}
\alias{readArraySDML}
\alias{readListSDML}
\alias{writeDescriptionSDML}
\alias{readDatasetSDML}
\alias{readProperties}
\alias{writeDfSDML}
\alias{readDescriptionSDML}
\alias{writePropertiesSDML}
\alias{getDataSDML}
\alias{getComplexDataSDML}
\alias{default}
\alias{readType}
\alias{writeTypeSDML}
\alias{tags}
\alias{getTextDataSDML}
\alias{cetagsSDML}
\alias{readCategoriesSDML}
\alias{writeCategoriesSDML}
\alias{writeListArraySDML}
\title{Internal StatDataML functions}
\description{
  Internal StatDataML functions.
}
\usage{}
\details{
  These are not to be called by the user.
}
\keyword{internal}
